Increase column width to handle all current option widths.
This should eventually be automatically calculated, with a maximum width
between the shortest option and its help text, a maximum absolute width,
and possibly a minimum required percentage of help to fit on one line.

This starts to get into some sort of heuristics, though. I think that
maybe we should just leave most of this as user-controlled. It's hard
for a program to have good formatting taste, and is outside the scope of
this requirement. Maybe it'll become worth it if this ever becomes a
standalone library.